refactor: 이전 handleApiResponse 제거
